Do not assume "page" parameters always exist
The API this application currently uses to query postcodes has a
pagination component in the response.

The API's expected response on the first page is:

```
{
    "count": 3768,
    "next": "https://staging.laalaa.dsd.io/legal-advisers/?page=2&postcode=BS20BL",
    "previous": null,
    "results": [<snip>
```

The API's expected response on the second page is:

```
{
    "count": 3768,
    "next": "https://staging.laalaa.dsd.io/legal-advisers/?page=3&postcode=BS20BL",
    "previous": "https://staging.laalaa.dsd.io/legal-advisers/?page=1&postcode=BS20BL",
    "results": [<snip>
```

If we upgrade the API application's dependencies then the `previous`
link will change by losing the `page=1` query parameter:

```diff
-"previous": "https://staging.laalaa.dsd.io/legal-advisers/?page=1&postcode=BS20BL",
+"previous": "https://staging.laalaa.dsd.io/legal-advisers/?postcode=BS20BL",
```

The removal of the `page` argument breaks the context filter, as it
makes the assumption that it will always be available:

```
File ".../fala/apps/adviser/templatetags/adviser_extras.py", line 25, in query_to_dict
    result = result[prop]
KeyError: 'page'
```

This changeset removes the assumption that "prop"s have to exist.