Expose uwsgi service on http protocol
In the template-deploy world the application runs in a Docker container
but the nginx service is running on the host machine.

We don't know how to reconfigure the nginx server running on
template-deploy to use the `uwsgi` protocol currently exposed, so I am
taking the route of reverting the configuration to use the `http`
protocol.

This will most likely **break** the Kubernetes configuration and it
will have to be fixed.